[Fix] 필기 폭을 Double로 저장

+++ semomun/StudyVCs/ChildVCs/FormTwo/MultipleWith5AnswerWideVC.swift
@@ -69,7 +69,7 @@ final class MultipleWith5AnswerWideVC: FormTwo {
     }
     
     override func updatePagePencilData(data: Data, width: CGFloat) {
-        self.viewModel?.updatePagePencilData(to: data, width: Int(width))
+        self.viewModel?.updatePagePencilData(to: data, width: Double(width))
     }
     
     override func previousPage() {
